# Description

This PR replaces the use of `Word32` with `SizeInBytes` for
`BlockComponents` and `SimpleStdHeader`, as per in #532.

+++ ouroboros-consensus/src/unstable-mock-block/Ouroboros/Consensus/Mock/Node/Serialisation.hs
@@ -72,7 +72,7 @@ instance Serialise ext => SerialiseNodeToNodeConstraints (MockBlock ext) where
       7 {- CBOR-in-CBOR -} + 1 {- encodeListLen 2 -} + hdrSize + bodySize
     where
       hdrSize  = fromIntegral (Lazy.length (serialise hdr))
-      bodySize = fromIntegral (simpleBodySize (simpleHeaderStd hdr))
+      bodySize = simpleBodySize (simpleHeaderStd hdr)
 
 instance Serialise ext => SerialiseNodeToNode (MockBlock ext) (MockBlock ext) where
   encodeNodeToNode _ _ = defaultEncodeCBORinCBOR
